requirements.txt
Flask==2.2.3
flask-cors==3.0.10
google-ai-generativelanguage==0.1.0
google-api-core==2.11.0
google-auth==2.17.3
google-generativeai==0.1.0rc1
googleapis-common-protos==1.59.0
